Refactor components structure
