+++ src/App.js
@@ -1,8 +1,8 @@
 import logo from './logo.svg';
 import './App.css';
 import MainPage from './components/Home-Page/MainPage';
-import NavBar from './components/Home-Page/NavBar';
-import Footer from './components/Home-Page/Footer';
+import NavBar from './components//NavBar';
+import Footer from './components/Footer';
 
 function App() {
   return (
